--- src/main/java/org/example/api/RetrofitClient.java~
@@ -1,4 +1,0 @@
-package org.example.api;
-
-public class RetrofitClient {
-}
