--- src/main/java/org/example/api/createupdatedelete/Response.java
@@ -1,48 +1,0 @@
-package org.example.api.createupdatedelete;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class Response {
-
-    @JsonProperty("createdAt")
-    private String createdAt;
-
-    @JsonProperty("updatedAt")
-    private String updatedAt;
-
-    @JsonProperty("id")
-    private int id;
-
-    @JsonProperty("name")
-    private String name;
-
-    @JsonProperty("username")
-    private String username;
-
-    @JsonProperty("email")
-    private String email;
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getCreatedAt() {
-        return createdAt;
-    }
-
-    public String getUpdatedAt() {
-        return updatedAt;
-    }
-}
