--- src/main/java/org/example/api/getById/ResponseUserGetByid.java~
@@ -1,4 +1,0 @@
-package org.example.api.getById;
-
-public class ResponseUserGetByid {
-}
